Don't wait on ContactsProvider
Right now during the initial WebSettings sync to native we wait
for the autofill profile to be loaded from disk so that it can
be synced. If there's no profile set, then we try to infer a
profile from the user's Me contact profile. Querying the Me
contact can be slow and in some extreme cases can cause the
settings sync on the UI thread to block long enough to show
an ANR.

Instead signal the threads (via the CountdownLatch) waiting on
the initial import before we do the Me profile lookup. Note that
we still may block the UI thread if the lookup of an already saved
profile takes an exceptionally long time. This is so that when a
user has saved a profile, we'll never resort to showing them the
"setup autofill" message. (But all ANR reports to date have shown
that we were querying the Me contact at the time of ANR).

Bug: 6371781
Change-Id: Ibb0d5e285ec3587d9f9bad3e69b79890850c2f6d