Remove WebSettings on destroy to avoid crash
When Tab.destroy is called there is a chance that the
WebSettings objects still exist in the mManagedSettings list.
If BrowserSettings.syncManagedSettings is called after
Tab.destroy and before automatic garbage collection
there will be a null pointer exception in ZoomManager.setZoomScale.
fix: WebSettings object connected to the WebView(Tab.java)
must be removed from the mManagedSettings list in
BrowserSettings.java when Tab.destroy is called.

Change-Id: Id542325ad53f161ed99175ea53214802be24834c