PhoneUI hardware keyboard support
Only XLargeUi(tablet) had support for some key events such as
KEYCODE_DPAD_UP to edit the url. But a hardware keyboard can be
connected to a phone so it should have the same behavior.

The code to handle key events has been moved from XLargeUi to
BaseUI, the common super-class for XLargeUi and PhoneUi.

Change-Id: I55998979ca52e2a10abcb3ba877a2533dd18be17