Encode characters that java.net.URI rejects
Bug: 5952386

Our java.net.URI implementation conforms to an old obsolete RFC and
it is very restrictive. Since in our download path, we use download
manager and so java.net.URI, this causes an odd bug (i.e. URIs that
are fetched fine by chromium http stack fails when download manager is
used). Also there is a second bug that when URI parsing fails and an
exception is thrown, we fail to catch it and crash. This CL fixes both.

Change-Id: I62ac289566efae97dd2161b8041b06a0a87211cb