Update patch set 1
Patch Set 1:

What device are you seeing this on? Catching the OOM is just ignoring the underlying problem (memory leak, somehow ending up with an absurdly large width/height, etc...) - according to the log the heap is only at 16647K which is well below what the max should be. As in, dalvik should have grown the heap instead of throwing an OOM.

Patch-set: 1
Label: Code-Review=0