Update patch set 1
Patch Set 1:

The real problem in that crash is the current memory usage which is listed as 45206K - that's waaaay too high. What device are you seeing this on and can you reproduce? There seems to be a memory leak which is what needs to be fixed instead of adding random OOM exception handlers.

Patch-set: 1