Update patch set 1
Patch Set 1: Do not submit

In that case I would opt to not take this patch. If you hit this again please file a bug, include a bug report (or at least more of logcat), and ideally get a memory dump so we can track down the leak itself.

Patch-set: 1
Label: Code-Review=-2