Update patch set 1
Patch Set 1: Code-Review-2

This means you got an onNewIntent either before onCreate or after onDestroy - in neither case should this create a real controller instance as the activity is not in a valid state.

If it stayed like this that is indicative of a bug in activity manager for whatever device you are running on.

Patch-set: 1
Reviewer: John Reck <1010103@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2