Update patch set 1
Patch Set 1:

Ah, I see. I think this may effect local-only bookmarks and not those that are synced, as the sync will then do a full cascade delete of bookmarks marked with IS_DELETED.

The fix should be in BrowserProvider2 where there is already some logic for this sort of thing, if you want to make the change I'll be happy to review it. However I will note that packages/apps/Browser is no longer being actively worked on as it was replaced with Chrome.

Patch-set: 1
Label: Code-Review=0