Update patch set 2
Patch Set 2:

Can't take this as is, it introduces race conditions between onCreate() and the rest of the ContentProvider methods.

With that said, this seems like trying to fix the wrong thing. For an ANR to occur there, either the file system or the disk is seriously hosed.

Patch-set: 2
Reviewer: John Reck <1010103@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0