Fix the problem that the layout of the suggestion drop down list does not fit to the screen width when an orientation changes.
Steps to reproduce:
1. Set the orientation of the Browser to landscape mode.
2. Type some keys in the url input view to show some suggestions.
3. Set the orientation to portrait mode again.

What heppended:
The width of the suggestion layout exceeds the screen width.
The arrow icon on the right side couldn't be seen.

Additional comments:
On some devices, setupDropDown() doesn't work correctly.
On the Samsung Galaxy S3, for instance, setupDropDown() produces left-hand margins.

The simplest way to fix these problems is to call setDropDownAnchor() when the url input view is created.

Change-Id: Id4d9273c6900bb1be4762e57fc489c00ab32268b
Signed-off-by: Sungmann Cho <sungmann.cho@navercorp.com>