Update patch set 1
Patch Set 1: Code-Review-1

This is working around the issue, not fixing the issue. The crash in libhwui is what needs to be fixed. The bug has been assigned to the webview team for further investigation as to what is causing the crash.

Patch-set: 1
Reviewer: John Reck <1010103@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1