Update patch set 3
Patch Set 3:

(2 comments)

Why are we continuing to allow file:// scheme URLs at all here? No intent filter matches that scheme any more, right? So the only way we will end up with one is if some app sends an explicit intent to the browser, which seems unlikely to be legitimate (the browser is not even present on all devices, or may not have the same package name, so sending explicit intents to it is not a reliable strategy to implement features).

Patch-set: 3
Reviewer: Richard Coles <1021004@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0