Create a new instance of intent instead of modify from getIntent directly
Activity#getIntent and ActivityThread#ActivitClientRecord#intent are the same
instance. If using getIntent and modify its component directly, the Record
of ActivityThread may become confused.

The token(IBinder)/activityInfo may be different from component of intent
after modifying, that can effect creating a new activity instance.
When AMS relaunching the token (maybe in monkey test), created activity is
not correct, and theme from activityInfo is wrong either.

So create a new Intent instance, and modify component to new instnace of intent
and send.

Change-Id: I7631b9f1d56dc66f0e9fa9b41bb1b2076b27bb09